+++ src/txio.h
@@ -205,6 +205,6 @@ public:
     TexOutStream();
     void               close(int chan);
     void               open(int chan, const std::string &file_name);
-    [[nodiscard]] auto is_open(int i) const -> bool { return write_open[i]; }
-    void               write(int chan, const std::string &s) { *(write_file[chan]) << s; }
+    [[nodiscard]] auto is_open(size_t i) const -> bool { return write_open[i]; }
+    void               write(size_t chan, const std::string &s) { *(write_file[chan]) << s; }
 };
